adding index.js / integration test
before it would not work because stylelint actually seems to need the module.export. The default export has been remove from the index.ts file in favour of an index.js file to maintain compatibility with the previous version of this package. An integration test with the actual stylelint has also been added.